sample/requirements.txt
Flask
python-dotenv
MyAPIManager
